將action從restart拆成 start stop restart

+++ Raspberry Pi/Final-Project/app/app.rb
@@ -8,14 +8,37 @@ Config.path(settings.config+'/config')
 get '/' do
   erb :index
 end
-get '/restart' do
-  if $pid && $ps
+get '/start' do
+  if $ps.nil? || $ps.closed?
+    $ps = IO.popen("rvmsudo ruby RYG.rb")
+    $pid = $ps.pid
+    @message = "start success"
+  else
+    @message = "is started already"
+  end
+  erb :led_action, :locals => {:message => @message}
+end
+get '/stop' do
+  if $pid && $ps && !$ps.closed?
     Process.kill("INT", $pid)
     $ps.close
+    @message = "stop success"
+  else
+    @message = "is stoped already"
   end
-  $ps = IO.popen("rvmsudo ruby RYG.rb")
-  $pid = $ps.pid
-  erb :restart
+  erb :led_action, :locals => {:message => @message}
+end
+get '/restart' do
+  if $pid && $ps && !$ps.closed?
+    Process.kill("INT", $pid)
+    $ps.close
+    $ps = IO.popen("rvmsudo ruby RYG.rb")
+    $pid = $ps.pid
+    @message = "restart success"
+  else
+    @message = "can't restart, maybe there is no traffic lights running"
+  end
+  erb :led_action, :locals => {:message => @message}
 end
 get '/config' do
   @config = Config.get
